Add basic support for B7 RC / KVMFR 20
This is a minor patch adding partial support for the updated KVMFR
version. Support for the RGB 24-in-32 operating mode may not work
as expected, but non-compressed modes used in previous versions,
such as BGRA/RGBA, should continue to work.

Note that the overarching plan for LGProxy is still to overhaul the
architecture, as this code is a mess. This is a stop-gap measure so
that users can benefit from the new Looking Glass features within
LGProxy.

Creation of further discussions or issues is welcome, so that new
feature additions and bugfixes can be triaged based on user feedback.

Changes:

- Use upstream Looking Glass again, as I won't maintain the texture
  compression fork
- Remove compressed texture support (see above). If you need texture
  compression support, use the old version `1ef3453`.
- Update copyright short identifier to GPL 2.0 or later
- Update help messages
- Use the OS name field in KVMFR to display the LGProxy relay version
- Update C standard to C11. It should already have been used in
  practice due to LG using it, but we define it here explicitly now.
- Enable optimization. We barely tested LGProxy with optimization, so
  it might break things.
- Enable sleep periods under 1ms and sync mode. This still has a quite
  a few problems that have to be ironed out, but in theory it will
  further reduce CPU usage.

Specific Recommendations:

- For TCP users, we recommend trying the new native TCP provider in
  Libfabric by updating your Libfabric versions (on both sides) to
  version 1.18.0 or higher. While this is no alternative to RDMA, it
  should provide improved performance in TCP mode.